add base_reidentification as base class for person and face reid
