+++ dynamic_vino_lib/include/dynamic_vino_lib/inferences/person_reidentification.hpp
@@ -25,6 +25,7 @@
 #include "dynamic_vino_lib/models/person_reidentification_model.hpp"
 #include "dynamic_vino_lib/engines/engine.hpp"
 #include "dynamic_vino_lib/inferences/base_inference.hpp"
+#include "dynamic_vino_lib/inferences/base_reidentification.hpp"
 #include "inference_engine.hpp"
 #include "opencv2/opencv.hpp"
 // namespace
@@ -100,24 +101,11 @@ public:
    * @return The name of the Inference instance.
    */
   const std::string getName() const override;
-  /**
-   * @brief Calculate the similarity between a new detected person with
-   * the already recorded persons.
-   * @return The similarity value.
-   */
-  float calcSimilarity(const std::vector<float> &, const std::vector<float> &);
-  /**
-   * @brief Try to find the matched person from the recorded persons, if there are not,
-   * record it in the recorded persons.
-   * @return The id of the matched person (or the new person).
-   */
-  std::string findMatchPerson(const std::vector<float> &);
 
 private:
   std::shared_ptr<Models::PersonReidentificationModel> valid_model_;
   std::vector<Result> results_;
-  std::vector<std::vector<float>> recorded_persons_;
-  double match_thresh_ = 0;
+  std::shared_ptr<dynamic_vino_lib::Tracker> person_tracker_;
 };
 }  // namespace dynamic_vino_lib
 #endif  // DYNAMIC_VINO_LIB__INFERENCES__PERSON_REIDENTIFICATION_HPP_
