+++ dynamic_vino_lib/src/inferences/person_reidentification.cpp
@@ -31,7 +31,10 @@ dynamic_vino_lib::PersonReidentificationResult::PersonReidentificationResult(
 
 // PersonReidentification
 dynamic_vino_lib::PersonReidentification::PersonReidentification(double match_thresh)
-: match_thresh_(match_thresh), dynamic_vino_lib::BaseInference() {}
+: dynamic_vino_lib::BaseInference()
+{
+  person_tracker_ = std::make_shared<dynamic_vino_lib::Tracker>(100000, match_thresh, 0.3);
+}
 
 dynamic_vino_lib::PersonReidentification::~PersonReidentification() = default;
 void dynamic_vino_lib::PersonReidentification::loadNetwork(
@@ -73,59 +76,13 @@ bool dynamic_vino_lib::PersonReidentification::fetchResults()
   for (int i = 0; i < getResultsLength(); i++) {
     std::vector<float> new_person = std::vector<float>(
       output_values + 256 * i, output_values + 256 * i + 256);
-    std::string person_id = findMatchPerson(new_person);
+    std::string person_id = "No." + std::to_string(
+      person_tracker_->processNewTrack(new_person));
     results_[i].person_id_ = person_id;
     found_result = true;
   }
   if (!found_result) {results_.clear();}
   return true;
-}
-
-float dynamic_vino_lib::PersonReidentification::calcSimilarity(
-  const std::vector<float> & person_a, const std::vector<float> & person_b)
-{
-  if (person_a.size() != person_b.size()) {
-    throw std::logic_error("cosine similarity can't be called for vectors of different lengths: "
-            "person_a size = " + std::to_string(person_a.size()) +
-            "person_b size = " + std::to_string(person_b.size()));
-  }
-  float mul_sum, denom_a, denom_b, value_a, value_b;
-  mul_sum = denom_a = denom_b = value_a = value_b = 0;
-  for (auto i = 0; i < person_a.size(); i++) {
-    value_a = person_a[i];
-    value_b = person_b[i];
-    mul_sum += value_a * value_b;
-    denom_a += value_a * value_a;
-    denom_b += value_b * value_b;
-  }
-  if (denom_a == 0 || denom_b == 0) {
-    throw std::logic_error("cosine similarity is not defined whenever one or both "
-            "input vectors are zero-vectors.");
-  }
-  return mul_sum / (sqrt(denom_a) * sqrt(denom_b));
-}
-
-std::string dynamic_vino_lib::PersonReidentification::findMatchPerson(
-  const std::vector<float> & new_person)
-{
-  auto size = recorded_persons_.size();
-  std::string id = "No.";
-  float best_match_sim = 0;
-  int best_match_ind = -1;
-  for (auto i = 0; i < size; ++i) {
-    float cos_sim = calcSimilarity(new_person, recorded_persons_[i]);
-    if (cos_sim > best_match_sim) {
-      best_match_sim = cos_sim;
-      best_match_ind = i;
-    }
-  }
-  if (best_match_sim > match_thresh_) {
-    recorded_persons_[best_match_ind] = new_person;
-    return id + std::to_string(best_match_ind);
-  } else {
-    recorded_persons_.push_back(new_person);
-    return id + std::to_string(size);
-  }
 }
 
 const int dynamic_vino_lib::PersonReidentification::getResultsLength() const
